pyproject.toml
[project]
name = "tackleberry"
version = "0.1.0"
description = "Tackleberry (or TB) is helping you tackle the access to AI"
authors = [
    { name = "Torsten Raudßus", email = "torsten@raudssus.de" },
]
dependencies = [
    "pydantic>=2.0.0",
    "instructor>=1.7.0",
    "typing-extensions>=4.0.0",
    "pyyaml>=6.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "flake8>=6.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "mock>=5.0.0",
    "requests",
    "openai",
    "ollama",
    "transformers",
    "groq",
    "instructor[groq]",
    "anthropic",
    "instructor[anthropic]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.hatch.build.targets.wheel]
packages = ["tackleberry"]

[tool.pytest]
testpaths = ["tests"]
python_files = "test_*.py"
